requirements.txt
dj-database-url==0.3.0
Django==1.7.5
django-admin-bootstrapped==2.3.4
django-dbarray==0.2
django-extensions==1.5.0
django-postgrespool==0.3.0
gunicorn==19.3.0
praw==2.1.20
psycopg2==2.6
requests==2.5.3
six==1.9.0
SQLAlchemy==0.9.8
update-checker==0.11
whitenoise==1.0.6
